+++ R/configuration.R
@@ -102,7 +102,7 @@
 #' defined locally for individual cassettes.
 #'
 #' - `record` (character) One of 'all', 'none', 'new_episodes', or 'once'.
-#' See [recording]
+#'   See [use_cassette()] for details.
 #' - `match_requests_on` vector of matchers. Default: (`method`, `uri`)
 #'   See [use_cassette()] for details.
 #' - `serialize_with`: (character) "yaml" or "json". Note that you can have
@@ -282,9 +282,9 @@ VCRConfig <- R6::R6Class(
     log_opts = function(value) {
       if (missing(value)) return(private$.log_opts)
       # add missing log options
-      log_opts <- merge_list(
-        value,
-        list(file = "vcr.log", log_prefix = "Cassette", date = TRUE)
+      log_opts <- utils::modifyList(
+        list(file = "vcr.log", log_prefix = "Cassette", date = TRUE),
+        value
       )
       private$.log_opts <- log_opts
     },
@@ -430,3 +430,12 @@ VCRConfig <- R6::R6Class(
 )
 
 pastec <- function(x) paste0(x, collapse = ", ")
+
+trimquotes <- function(x, y) {
+  pattern <- "^\"|\"$|^'|'$"
+  if (grepl(pattern, x)) {
+    msg <- "filter_sensitive_data: leading & trailing quotes trimmed from '"
+    warning(paste0(msg, y, "'"), call. = FALSE)
+  }
+  gsub(pattern, "", x)
+}