+++ R/request_handler-httr.R
@@ -1,25 +1,17 @@
-#' @title RequestHandlerHttr
-#' @description Methods for the httr package, building on [RequestHandler]
 #' @export
-#' @param request A request object
 RequestHandlerHttr <- R6::R6Class(
   "RequestHandlerHttr",
   inherit = RequestHandler,
 
   public = list(
-    #' @description Create a new `RequestHandlerHttr` object
-    #' @param request A request object
-    #' @return A new `RequestHandlerHttr` object
     initialize = function(request) {
       self$request_original <- request
-      self$request <- {
-        Request$new(
-          request$method,
-          request$url,
-          curl_body(request),
-          as.list(request$headers)
-        )
-      }
+      self$request <- vcr_request(
+        request$method,
+        request$url,
+        curl_body(request),
+        as.list(request$headers)
+      )
     }
   ),
 
@@ -78,12 +70,6 @@ RequestHandlerHttr <- R6::R6Class(
   )
 )
 
-disk_true <- function(x) {
-  if (is.null(x)) return(FALSE)
-  assert(x, "logical")
-  return(x)
-}
-
 # generate actual httr response
 serialize_to_httr <- function(request, response) {
   # request
@@ -107,17 +93,13 @@ serialize_to_httr <- function(request, response) {
   # in vcr >= v0.4, "disk" is in the response, but in older versions
   # its missing - use response$body if disk is not present
   response_body <- response$body
-  if ("disk" %in% names(response) && disk_true(response$disk)) {
-    response_body <- if (response$disk) {
-      structure(response$body, class = "path")
-    } else {
-      response$body
-    }
+  if (response$disk) {
+    response_body <- structure(response$body, class = "path")
   }
-  resp$set_body(response_body, response$disk %||% FALSE)
+  resp$set_body(response_body, response$disk)
   resp$set_request_headers(request$headers, capitalize = FALSE)
   resp$set_response_headers(response$headers, capitalize = FALSE)
-  resp$set_status(status = response$status$status_code %||% 200)
+  resp$set_status(status = response$status)
 
   # generate httr response
   webmockr::build_httr_response(as_httr_request(req), resp)