+++ R/request_handler-httr2.R
@@ -1,28 +1,20 @@
-#' @title RequestHandlerHttr2
-#' @description Methods for the httr2 package, building on [RequestHandler]
 #' @export
-#' @param request The request from.
 RequestHandlerHttr2 <- R6::R6Class(
   "RequestHandlerHttr2",
   inherit = RequestHandler,
 
   public = list(
-    #' @description Create a new `RequestHandlerHttr2` object
-    #' @param request The request
-    #' @return A new `RequestHandlerHttr2` object
     initialize = function(request) {
       if (!length(request$method)) {
         request$method <- webmockr:::req_method_get_w(request)
       }
       self$request_original <- request
-      self$request <- {
-        Request$new(
-          request$method,
-          request$url,
-          httr2_body(request),
-          request$headers
-        )
-      }
+      self$request <- vcr_request(
+        request$method,
+        request$url,
+        httr2_body(request),
+        request$headers
+      )
     }
   ),
 
@@ -44,7 +36,7 @@ RequestHandlerHttr2 <- R6::R6Class(
       }
 
       httr2::response(
-        status_code = vcr_response$status$status_code %||% 200,
+        status_code = vcr_response$status,
         url = self$request_original$url,
         method = webmockr:::req_method_get_w(self$request_original),
         headers = vcr_response$headers,