+++ tests/testthat/test-errors.R
@@ -1,141 +1,51 @@
 test_that("UnhandledHTTPRequestError fails well", {
-  request <- Request$new("post", hb("/post?a=5"), "", list(foo = "bar"))
+  local_cassette("test", warn_on_empty = FALSE)
+  expect_snapshot(UnhandledHTTPRequestError$new(5), error = TRUE)
+})
 
-  z <- UnhandledHTTPRequestError$new(request)
-  expect_error(
-    z$construct_message(),
-    "There is currently no cassette in use"
-  )
+test_that("informative error if no cassette active", {
+  withr::local_options(rlang_interactive = TRUE)
 
-  local_cassette("asdsfd", warn_on_empty = FALSE)
-  # # types
-  expect_error(
-    UnhandledHTTPRequestError$new(5),
-    "request must be of class Request"
-  )
+  request <- vcr_request("post", "http://example.com")
+  err <- UnhandledHTTPRequestError$new(request)
+  expect_snapshot(err$construct_message(), error = TRUE)
 })
 
 test_that("UnhandledHTTPRequestError works as expected", {
-  request <- Request$new("post", hb("/post?a=5"), "", list(foo = "bar"))
+  withr::local_options(rlang_interactive = TRUE)
   local_cassette("turtle", warn_on_empty = FALSE)
 
-  a <- UnhandledHTTPRequestError$new(request)
-
-  expect_s3_class(a, "UnhandledHTTPRequestError")
-  expect_s3_class(a$cassette, "Cassette")
-  expect_equal(a$cassette$name, "turtle")
-  expect_type(a$construct_message, "closure")
-
-  expect_error(
-    a$construct_message(),
-    "An HTTP request has been made that vcr does not know how to handle"
-  )
+  request <- vcr_request("post", "http://example.com")
+  err <- UnhandledHTTPRequestError$new(request)
+  expect_snapshot(err$construct_message(), error = TRUE)
 })
 
-test_that("UnhandledHTTPRequestError works as expected", {
-  ## API key in query param
-  withr::local_envvar(
-    FOO_BAR = "2k2k2k288gjrj2i21i",
-    HELLO_WORLD = "asdfadfasfsfs239823n23"
-  )
-  local_vcr_configure(
-    filter_sensitive_data = list(
-      "<<foo_bar_key>>" = Sys.getenv("FOO_BAR"),
-      "<<hello_world_key>>" = Sys.getenv("HELLO_WORLD")
-    )
-  )
-  url <- paste0(
-    hb("/get?api_key="),
-    Sys.getenv("FOO_BAR"),
-    "&other_secret=",
-    Sys.getenv("HELLO_WORLD")
-  )
-  request <- Request$new("get", url, "")
+test_that("sensitive data is redacted in url", {
   local_cassette("bunny", warn_on_empty = FALSE)
 
-  a <- UnhandledHTTPRequestError$new(request)
+  foo <- "abc"
+  local_vcr_configure(filter_sensitive_data = list("<<foo>" = foo))
+  url <- paste0("http://example.com/?foo=", foo)
+  request <- vcr_request("get", url)
 
-  expect_s3_class(a, "UnhandledHTTPRequestError")
-  expect_s3_class(a$cassette, "Cassette")
-  expect_equal(a$cassette$name, "bunny")
-  expect_type(a$construct_message, "closure")
-
-  expect_error(
-    a$construct_message(),
-    "An HTTP request has been made that vcr does not know how to handle"
-  )
-  expect_error(
-    a$construct_message(),
-    "<<foo_bar_key>>"
-  )
-  expect_error(
-    a$construct_message(),
-    "<<hello_world_key>>"
-  )
+  err <- UnhandledHTTPRequestError$new(request)
+  cnd <- catch_cnd(err$construct_message(), classes = "error")
+  expect_no_match(conditionMessage(cnd), foo)
 })
 
-test_that("UnhandledHTTPRequestError works as expected", {
-  ## API key in header
-  withr::local_envvar(FOO_BAR = "2k2k2k288gjrj2i21i")
-  local_vcr_configure(
-    dir = withr::local_tempdir(),
-    filter_sensitive_data = list("<<foo_bar_key>>" = Sys.getenv("FOO_BAR")),
-  )
-  url <- hb("/get")
-  request <- Request$new("get", url, "", list(api_key = Sys.getenv("FOO_BAR")))
+test_that("sensitive data is redacted in header", {
   local_cassette(
-    "frog",
+    "test",
     match_requests_on = c("method", "uri", "headers"),
     warn_on_empty = FALSE
   )
 
-  a <- UnhandledHTTPRequestError$new(request)
+  foo <- "abc"
+  local_vcr_configure(filter_sensitive_data = list("<<foo>>" = foo))
+  url <- "http://example.com"
+  request <- vcr_request("get", url, "", list(api_key = foo))
 
-  expect_s3_class(a, "UnhandledHTTPRequestError")
-  expect_s3_class(a$cassette, "Cassette")
-  expect_equal(a$cassette$name, "frog")
-  expect_type(a$construct_message, "closure")
-
-  expect_error(
-    a$construct_message(),
-    "An HTTP request has been made that vcr does not know how to handle"
-  )
-  expect_error(
-    a$construct_message(),
-    "api_key: <<foo_bar_key>>"
-  )
-})
-
-test_that("UnhandledHTTPRequestError works as expected", {
-  ## API key not found or empty (i.e., "")
-  withr::local_envvar(HELLO_MARS = "asdfadfasfsfs239823n23")
-  local_vcr_configure(
-    filter_sensitive_data = list(
-      "<<bar_foo_key>>" = Sys.getenv("BAR_FOO"),
-      "<<hello_mars_key>>" = Sys.getenv("HELLO_MARS")
-    )
-  )
-  url <- paste0(hb("/get?api_key="), Sys.getenv("HELLO_MARS"))
-  request <- Request$new("get", url, "")
-  local_cassette("bunny2", warn_on_empty = FALSE)
-
-  a <- UnhandledHTTPRequestError$new(request)
-
-  expect_s3_class(a, "UnhandledHTTPRequestError")
-  expect_s3_class(a$cassette, "Cassette")
-  expect_equal(a$cassette$name, "bunny2")
-  expect_type(a$construct_message, "closure")
-
-  expect_error(
-    a$construct_message(),
-    "An HTTP request has been made that vcr does not know how to handle"
-  )
-  # we should see this string in the error message
-  expect_error(
-    a$construct_message(),
-    "<<hello_mars_key>>"
-  )
-  # we should not see this string in the error message b/c it's empty/missing
-  res <- tryCatch(a$construct_message(), error = function(e) e)
-  expect_false(grepl("<<bar_foo_key>>", res$message))
+  err <- UnhandledHTTPRequestError$new(request)
+  cnd <- catch_cnd(err$construct_message(), classes = "error")
+  expect_no_match(conditionMessage(cnd), foo)
 })